device_lights: Remove __packed from struct channel_config
The __packed macro was not defined, resulting in it being treated
as a variable name. This used additional RAM as a "__packed"
symbol was added to memory, as seen in the System.map file.
The __packed macro is defined in a greybus header file, which is
not appropriate to include here. Switching to __attribute__((packed))
increases the code size by several bytes, so that option was not
chosen. Because the structure was not actually packed before, and
because it is not being sent outside of the uC, the __packed was
simply removed completely.

Change-Id: I7a2607e3251504bc96572c5be756eefea605ba88
Signed-off-by: Greg Meiste <w30289@motorola.com>
Reviewed-on: http://gerrit.mot.com/857611
SME-Granted: SME Approvals Granted
Submit-Approved: Jira Key <jirakey@motorola.com>
Tested-by: Jira Key <jirakey@motorola.com>
SLTApproved: Slta Waiver <sltawvr@motorola.com>
Reviewed-by: Michael Corrigan <michael.corrigan@motorola.com>
Reviewed-by: James Wylder <jwylder@motorola.com>