power: battery_level: Report full battery after it is fully charged
Wait for the battery to become fully charged to report that it is full
instead of checking that the battery level is above some threshold. We
are relying on a low level driver to report that battery level has
reached 100% if and only if the battery has become fully charged.

Change-Id: I5ee2d5e3ab92730eeb2a2e7cf4a4c84c47ca58fa
Signed-off-by: Dmitriy Gruzman <dgruzman@motorola.com>
Reviewed-on: http://gerrit.mot.com/861968
SME-Granted: SME Approvals Granted
Submit-Approved: Jira Key <jirakey@motorola.com>
Tested-by: Jira Key <jirakey@motorola.com>
SLTApproved: Slta Waiver <sltawvr@motorola.com>
Reviewed-by: Weiqi Tang <woodtang@motorola.com>
Reviewed-by: Gregory Meiste <w30289@motorola.com>