power: bq24292/bq25896: Filter out momentary boost faults
When switching the charge direction, momentary boost faults are
seen. Previously, this boost fault was immediately reported to the
system, which resulting in the charge current being shut off.

With this patch, a boost fault needs to be seen for 100ms before
it is reported. If greater than 100ms, it is likely to be a real
fault.

Unfortunately, there is no interrupt generated when the boost
fault condition is cleared, so the usual debouncing method will
not work.

Change-Id: I4f676782b901014c0f17ff4ab22d661ff2c8cdf8
Signed-off-by: Greg Meiste <w30289@motorola.com>
Reviewed-on: http://gerrit.mot.com/864577
SME-Granted: SME Approvals Granted
Reviewed-by: Dmitriy Gruzman <dgruzman@motorola.com>
Submit-Approved: Jira Key <jirakey@motorola.com>
Tested-by: Jira Key <jirakey@motorola.com>
SLTApproved: Slta Waiver <sltawvr@motorola.com>
Reviewed-by: Weiqi Tang <woodtang@motorola.com>