arm: stm32: spi: CRC can be enabled on individual port(s)
Previously, when CRC16 support was enabled, the driver was written
so all ports required CRC. However, it is likely that CRC is only
needed on specific port(s) and not others. This patch adds an
additional flag for specifying which port(s) need CRC enabled.
This flag must be defined in the board.h for the configs which
enable the CRC16 driver support.

Change-Id: Id0f801fe9ad146655d39b0ad3aaca061aee1ec0d
Signed-off-by: Greg Meiste <w30289@motorola.com>
Reviewed-on: http://gerrit.mot.com/869170
SME-Granted: SME Approvals Granted
Submit-Approved: Jira Key <jirakey@motorola.com>
Tested-by: Jira Key <jirakey@motorola.com>
SLTApproved: Slta Waiver <sltawvr@motorola.com>
Reviewed-by: James Wylder <jwylder@motorola.com>