configs: hdk: muc: Provide missing funcs for SPI1 and disable CRC
Since SPI1 is a master port routed to the 80 pin connector port
for pcard use, add the missing select and status functions for
it. This will allow the chip select to be toggled for communication.
The status function always returns zero for now and can be modified
to fit your needs of your pcard.

In addition, CRC is disabled on SPI1. Previously the low level SPI
driver was enabling the CRC functionality on all SPI ports when the
CRC support was enabled. The CRC is still required on SPI2 for
communication with the base, but odds are the hardware you chose
for your pcards will not require or even support CRC, so disable
CRC on SPI1.

Change-Id: Ibffccf9176c28b5eb9812b2219ce0820b4b8ed1d
Signed-off-by: Greg Meiste <w30289@motorola.com>
Reviewed-on: http://gerrit.mot.com/869172
SLTApproved: Slta Waiver <sltawvr@motorola.com>
SME-Granted: SME Approvals Granted
Submit-Approved: Jira Key <jirakey@motorola.com>
Tested-by: Jira Key <jirakey@motorola.com>
Reviewed-by: James Wylder <jwylder@motorola.com>