drivers: Kconfig: HDMI_DISPLAY: add configs to limit HDMI bandwidth
Add a pair of configs which enable bandwidth limiting on an HDMI display
and set the bandwidth limit to a specified value in KHz.

Add a new config option CONFIG_HDMI_DISPLAY_MAX_BANDWIDTH_KHZ which
limits the bandwidth that the host should try to use with a downstream
device. This config should be set to a positive integer which represents
the maximum allowed frequency in KHz. The host must enforce this limit
when enumerating supported display resolutions to ensure that signal
quality is high enough for the display link to work properly.

The option HDMI_DISPLAY_BANDWIDTH_LIMIT controls whether the max
bandwidth limit should be used or not.

Change-Id: I025d70de1b9d603197a08d6505c5ace05ebceda8
Reviewed-on: http://gerrit.mot.com/873077
SLTApproved: Slta Waiver <sltawvr@motorola.com>
SME-Granted: SME Approvals Granted
Submit-Approved: Jira Key <jirakey@motorola.com>
Tested-by: Jira Key <jirakey@motorola.com>
Reviewed-by: Patrick Auchter <auchter@motorola.com>
Reviewed-by: Michael Corrigan <michael.corrigan@motorola.com>
Reviewed-by: James Wylder <jwylder@motorola.com>
Reviewed-by: Gregory Meiste <w30289@motorola.com>