Revert "battery-ptp: Allow the Base to power off when Mod attached"
Revert the change that added the check for the current direction to be
"Off" before setting the flag to prevent applying battery power to the
powered off Core. Now, the battery power is not applied to the Core that
just turned off regardless of what the current direction was.

This is done to comply with the requirement stating that if the Core is
turned off, the Mod shall *NOT* apply power to the Core unless it has
an external power source.

This reverts commit c0c1073d0f0e2f4269d9c742c3d11d00432dd874.

Change-Id: Ibe64d231166995d4867f8d1dc5824deabfbb1670
Reviewed-on: http://gerrit.mot.com/871865
SME-Granted: SME Approvals Granted
Submit-Approved: Jira Key <jirakey@motorola.com>
Tested-by: Jira Key <jirakey@motorola.com>
SLTApproved: Slta Waiver <sltawvr@motorola.com>
Reviewed-by: James Wylder <jwylder@motorola.com>
Reviewed-by: Min Shi <e13386@motorola.com>
Reviewed-by: Gregory Meiste <w30289@motorola.com>