greybus: ptp: Support variable voltage charging
Add commands to support Mod to Phone and Phone to Mod variable voltage
charging.

For the Mod to Phone charging, the Phone will tell the Mod its max
input voltage and query the output voltage of the charging source.

For the Phone to Mod charging, the Phone will query the Mod max input
voltage and will tell the Mod the output voltage of the charging source.

Allow the Phone to query the current_flow direction since the direction
transitions to "off" whenever the voltage/current/type of the power
source supplying power to the Phone change.

Update protocol version to 00.03.

Change-Id: I7464f6a21b4a68ceb605f38d61ab8650ec5323e2
Signed-off-by: Dmitriy Gruzman <dgruzman@motorola.com>