tsb: vidcrypt: Fix vidcrypt for simultaneous CSI and DSI
The vidcrypt initialization logic had an issue where it always
reset the vidcrypt block whenever either CSI or DSI was started.
This broke the case where both CSI and DSI were started.

Add a reference count to the vidcrypt initialization to
initialize the core vidcrypt block the first time only.

Change-Id: I56fdb67d52e1e71e1c8187e7e89f2a943ee7f67c
Signed-off-by: Mike Corrigan <corrigan@motorola.com>