drivers: sensors: bme680: Add BME680 Greybus Sensors Ext driver
Add a BME680 driver that exposes the four sensors (temperature,
pressure, humidity, and gas) via the Greybus Sensors Ext interface.
The sensors appear in Android's SensorManager as four dynamic
sensors.

The driver is instaniated four separate times (once for each
sensor type) and it uses a device resource to identify the exact
flavor of sensor to activate.

Internally the driver uses a 'sub-sensor' concept to commonize
the logic between the four sensors and to allow all four sensors
to share a single polling thread.

The underlying I2C driver is based on Bosch Sensortec's
open-source and BSD-like licensed implementation found at
https://github.com/BoschSensortec/BME680_driver.

Signed-off-by: Mike Corrigan <corrigan@motorola.com>